[FIX] 요청 이미지와 현재 유저 이미지가 다를 경우만 기존 이미지 삭제 처리

+++ src/main/java/kr/hs/entrydsm/husky/domain/user/service/info/UserInfoServiceImpl.java
@@ -50,7 +50,7 @@ public class UserInfoServiceImpl implements UserInfoService {
         User user = userRepository.findById(receiptCode)
                 .orElseThrow(UserNotFoundException::new);
 
-        if (!user.isPhotoEmpty() && isExists(request.getPhoto()))
+        if (!user.isPhotoEmpty() && isExists(request.getPhoto()) && !user.getUserPhoto().equals(request.getPhoto()))
             imageService.delete(user.getUserPhoto());
 
         user.updateInfo(request);
